requirements.txt
cffi==1.15.0
click==8.0.3
Flask==2.0.2
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pycparser==2.21
PyNaCl==1.4.0
six==1.16.0
typing-extensions==4.0.1
Werkzeug==2.0.2
zipp==3.6.0
